Issue #154 refactor: LoginDataSource
