+++ app/src/main/java/com/lateinit/rightweight/data/datasource/LoginDataSource.kt
@@ -5,5 +5,6 @@ import com.lateinit.rightweight.data.LoginResponse
 interface LoginDataSource {
 
     suspend fun login(key: String, token: String): LoginResponse
+
     suspend fun deleteAccount(key: String, idToken: String)
 }
