+++ app/src/main/java/com/lateinit/rightweight/data/datasource/LoginDataSourceImpl.kt
@@ -1,7 +1,9 @@
 package com.lateinit.rightweight.data.datasource
 
-import com.lateinit.rightweight.data.LoginResponse
 import com.lateinit.rightweight.data.AuthApiService
+import com.lateinit.rightweight.data.LoginResponse
+import com.lateinit.rightweight.data.model.LoginRequestBody
+import com.lateinit.rightweight.data.model.PostBody
 import javax.inject.Inject
 
 class LoginDataSourceImpl @Inject constructor(
@@ -10,28 +12,16 @@ class LoginDataSourceImpl @Inject constructor(
 
     override suspend fun login(key: String, token: String): LoginResponse {
         val postBody = PostBody(token, "google.com").toString()
-        val loginResponse =
-            api.loginToFirebase(key, LoginRequestBody(postBody, "http://localhost", true, true))
-        return loginResponse
+        return api.loginToFirebase(
+            key, LoginRequestBody(
+                postBody, "http://localhost",
+                returnIdpCredential = true,
+                returnSecureToken = true
+            )
+        )
     }
 
     override suspend fun deleteAccount(key: String, idToken: String) {
         api.deleteAccount(key, idToken)
-    }
-}
-
-data class LoginRequestBody(
-    val postBody: String,
-    val requestUri: String,
-    val returnIdpCredential: Boolean,
-    val returnSecureToken: Boolean
-)
-
-data class PostBody(
-    val id_token: String,
-    val providerId: String,
-) {
-    override fun toString(): String {
-        return "id_token=$id_token&providerId=$providerId"
     }
 }
